+++ src/CachedBuilder.php
@@ -12,7 +12,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function avg($column)
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::avg($column);
         }
 
@@ -25,7 +25,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function count($columns = ['*'])
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::count($columns);
         }
 
@@ -38,7 +38,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function cursor()
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return collect(parent::cursor());
         }
 
@@ -62,12 +62,12 @@ class CachedBuilder extends EloquentBuilder
      */
     public function find($id, $columns = ['*'])
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::find($id, $columns);
         }
 
         $arguments = func_get_args();
-        $cacheKey = $this->makeCacheKey($columns);
+        $cacheKey = $this->makeCacheKey(['*'], null, "-find_{$id}");
         $method = 'find';
 
         return $this->cachedValue($arguments, $cacheKey, $method);
@@ -75,7 +75,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function first($columns = ['*'])
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::first($columns);
         }
 
@@ -88,7 +88,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function get($columns = ['*'])
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::get($columns);
         }
 
@@ -101,7 +101,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function max($column)
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::max($column);
         }
 
@@ -114,7 +114,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function min($column)
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::min($column);
         }
 
@@ -125,9 +125,27 @@ class CachedBuilder extends EloquentBuilder
         return $this->cachedValue($arguments, $cacheKey, $method);
     }
 
+    public function paginate(
+        $perPage = null,
+        $columns = ['*'],
+        $pageName = 'page',
+        $page = null
+    ) {
+        if (! $this->isCachable()) {
+            return parent::paginate($perPage, $columns, $pageName, $page);
+        }
+
+        $arguments = func_get_args();
+        $page = $page ?: 1;
+        $cacheKey = $this->makeCacheKey($columns, null, "-paginate_by_{$perPage}_{$pageName}_{$page}");
+        $method = 'paginate';
+
+        return $this->cachedValue($arguments, $cacheKey, $method);
+    }
+
     public function pluck($column, $key = null)
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::pluck($column, $key);
         }
 
@@ -141,7 +159,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function sum($column)
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::sum($column);
         }
 
@@ -154,7 +172,7 @@ class CachedBuilder extends EloquentBuilder
 
     public function value($column)
     {
-        if (! $this->isCachable) {
+        if (! $this->isCachable()) {
             return parent::value($column);
         }
 
@@ -197,7 +215,9 @@ class CachedBuilder extends EloquentBuilder
         string $method
     ) {
         if ($result['key'] !== $cacheKey) {
-            cache()->tags($cacheTags)->forget($hashedCacheKey);
+            $this->cache()
+                ->tags($cacheTags)
+                ->forget($hashedCacheKey);
 
             $result = $this->retrieveCachedValue(
                 $arguments,