+++ tests/Unit/Traits/CachableTest.php
@@ -19,37 +19,11 @@ class CachableTest extends UnitTestCase
 {
     use RefreshDatabase;
 
-    public function setUp()
-    {
-        parent::setUp();
-
-        cache()->flush();
-        $publishers = factory(Publisher::class, 10)->create();
-        factory(Author::class, 10)->create()
-            ->each(function ($author) use ($publishers) {
-                factory(Book::class, random_int(2, 10))->make()
-                    ->each(function ($book) use ($author, $publishers) {
-                        $book->author()->associate($author);
-                        $book->publisher()->associate($publishers[rand(0, 9)]);
-                        $book->save();
-                    });
-                factory(Profile::class)->make([
-                    'author_id' => $author->id,
-                ]);
-            });
-
-        $bookIds = (new Book)->all()->pluck('id');
-        factory(Store::class, 10)->create()
-            ->each(function ($store) use ($bookIds) {
-                $store->books()->sync(rand($bookIds->min(), $bookIds->max()));
-            });
-        cache()->flush();
-    }
-
     public function testSpecifyingAlternateCacheDriver()
     {
         $configCacheStores = config('cache.stores');
         $configCacheStores['customCache'] = ['driver' => 'array'];
+        // TODO: make sure the alternate cache is actually loaded
         config(['cache.stores' => $configCacheStores]);
         config(['laravel-model-caching.store' => 'customCache']);
         $key = sha1('genealabs:laravel-model-caching:genealabslaravelmodelcachingtestsfixturesauthor');
@@ -76,7 +50,8 @@ class CachableTest extends UnitTestCase
     {
         (new PrefixedAuthor)->get();
 
-        $results = cache()
+        $results = $this->
+            cache()
             ->tags([
                 'genealabs:laravel-model-caching:test-prefix:genealabslaravelmodelcachingtestsfixturesprefixedauthor',
             ])
@@ -85,21 +60,41 @@ class CachableTest extends UnitTestCase
         $this->assertNotNull($results);
     }
 
-    public function testAllReturnsCollection()
-    {
-        (new Author)->truncate();
-        factory(Author::class, 1)->create();
-        $authors = (new Author)->all();
+    // public function testAllReturnsCollection()
+    // {
+    //     (new Author)->truncate();
+    //     factory(Author::class, 1)->create();
+    //     $authors = (new Author)->all();
+    //
+    //     $cachedResults = $this
+    //         ->cache()
+    //         ->tags([
+    //             'genealabs:laravel-model-caching:genealabslaravelmodelcachingtestsfixturesauthor',
+    //         ])
+    //         ->get(sha1('genealabs:laravel-model-caching:genealabslaravelmodelcachingtestsfixturesauthor'))['value'];
+    //     $liveResults = (new UncachedAuthor)->all();
+    //
+    //     $this->assertInstanceOf(Collection::class, $authors);
+    //     $this->assertInstanceOf(Collection::class, $cachedResults);
+    //     $this->assertInstanceOf(Collection::class, $liveResults);
+    // }
 
-        $cachedResults = cache()
+    public function testsCacheFlagDisablesCaching()
+    {
+        config(['laravel-model-caching.disabled' => true]);
+
+        $authors = (new Author)->get();
+        $cachedAuthors = $this
+            ->cache()
             ->tags([
                 'genealabs:laravel-model-caching:genealabslaravelmodelcachingtestsfixturesauthor',
             ])
-            ->get(sha1('genealabs:laravel-model-caching:genealabslaravelmodelcachingtestsfixturesauthor'))['value'];
-        $liveResults = (new UncachedAuthor)->all();
+            ->get(sha1('genealabs:laravel-model-caching:genealabslaravelmodelcachingtestsfixturesauthor'));
 
-        $this->assertInstanceOf(Collection::class, $authors);
-        $this->assertInstanceOf(Collection::class, $cachedResults);
-        $this->assertInstanceOf(Collection::class, $liveResults);
+        config(['laravel-model-caching.disabled' => false]);
+
+        $this->assertNull($cachedAuthors);
+        $this->assertNotEmpty($authors);
+        $this->assertCount(10, $authors);
     }
 }